:hammer: refactor: package 명 수정
