+++ src/main/java/com/example/harumeonglog/domain/member/service/AppleOAuth2Service.java
@@ -2,7 +2,7 @@ package com.example.harumeonglog.domain.member.service;
 
 import com.example.harumeonglog.domain.member.dto.request.OAuth2Request;
 import com.example.harumeonglog.domain.member.dto.response.OAuth2Response;
-import com.example.harumeonglog.domain.member.infrastructure.MemberRepository;
+import com.example.harumeonglog.domain.member.repository.MemberRepository;
 import com.example.harumeonglog.global.error.exception.AuthException;
 import com.example.harumeonglog.global.util.WebClientUtil;
 import io.jsonwebtoken.Claims;
