+++ src/main/java/com/example/harumeonglog/domain/member/service/KakaoOAuth2Service.java
@@ -4,7 +4,7 @@ import com.example.harumeonglog.domain.member.converter.OAuth2Converter;
 import com.example.harumeonglog.domain.member.dto.request.OAuth2Request;
 import com.example.harumeonglog.domain.member.dto.response.OAuth2Response;
 import com.example.harumeonglog.domain.member.entity.enums.SocialType;
-import com.example.harumeonglog.domain.member.infrastructure.MemberRepository;
+import com.example.harumeonglog.domain.member.repository.MemberRepository;
 import com.example.harumeonglog.global.data.KakaoOAuthConfigData;
 import com.example.harumeonglog.global.error.exception.AuthException;
 import com.example.harumeonglog.global.util.WebClientUtil;
