+++ src/main/java/com/example/harumeonglog/global/security/service/CustomDetailService.java
@@ -4,7 +4,7 @@ import com.example.harumeonglog.domain.member.entity.Member;
 import com.example.harumeonglog.global.security.domain.CustomUserDetails;
 import com.example.harumeonglog.global.error.code.MemberErrorCode;
 import com.example.harumeonglog.global.error.exception.MemberException;
-import com.example.harumeonglog.domain.member.infrastructure.MemberRepository;
+import com.example.harumeonglog.domain.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
